# perf(simd): avx2 fallback to swar instead of sse4.2