@AaronO 2024-09-02
@seanmonstar @lucab I'll bench this against for #175 for completeness sake, but this is substantially simpler/focused, should not regress `arch64`/etc... and should be faster overall.